% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_cov.R
\name{cov}
\alias{cov}
\alias{within_cov}
\alias{between_cov}
\title{Calculate the uncorrected within and between covariance.}
\usage{
within_cov(within, between, id)

between_cov(between, id)
}
\arguments{
\item{within}{the data on within level including an id}

\item{between}{the data on the between level optionally including an id}

\item{id}{the name of the variable that serves as an identifier}
}
\description{
Calculate the uncorrected within and between covariance.
}
