% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_scores.R
\name{fs}
\alias{fs}
\alias{fs_within}
\alias{fs_between}
\title{Calculate factor scores}
\usage{
fs_within(models, data, ...)

fs_between(models, data, ...)
}
\arguments{
\item{models}{fitted multilevel lavaan models e.g. output of \code{\link{fit_models}}}

\item{data}{the dataset on which the models where fitted, rows which contain missings on the variables within models or ... are dropped.}

\item{...}{additional variables that should be retained}
}
\description{
Wrapper for \code{\link[lavaan:lavPredict]{lavaan::lavPredict}}.for multiple models.
}
\seealso{
\link{fsm}
}
