% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_scores.R
\name{fsm}
\alias{fsm}
\alias{fsm_within}
\alias{fsm_between}
\title{Calculate Factor Score Matrices}
\usage{
fsm_within(models, data, ...)

fsm_between(models, data, ...)
}
\arguments{
\item{models}{fitted multilevel lavaan models e.g. output of \code{\link{fit_models}}}

\item{data}{the dataset on which the models where fitted, rows which contain missings on the variables within models or ... are dropped.}

\item{...}{additional variables that should be retained}
}
\description{
Wrapper for \code{\link[lavaan:lavPredict]{lavaan::lavPredict}} (with \code{fsm = TRUE}) for multiple models.
}
\seealso{
\link{fs}
}
