% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_data}
\alias{simulate_data}
\alias{simulate_data_l2}
\alias{simulate_data_l1}
\title{Simulate Multilevel Data}
\usage{
simulate_data(model, nobs.l1, nobs.l2, ...)

simulate_data_l2(model.l2, sample.nobs, id, ...)

simulate_data_l1(model.l1, sample.nobs, data.l2, ...)
}
\arguments{
\item{model}{a two-level sem (first level one, then level two). Anything that \code{\link[lavaan:model.syntax]{lavaan::lavaanify}} accepts.}

\item{nobs.l1}{Number of observations on level 1. May be a vector of length nobs.l2}

\item{nobs.l2}{Number of observations on level 2.}

\item{...}{passed to \code{\link[lavaan:simulateData]{lavaan::simulateData}}}

\item{model.l2}{a simple model only for level 2}

\item{sample.nobs}{Number of observations.}

\item{id}{the cluster identifier}

\item{model.l1}{a simple model only for level 1.}

\item{data.l2}{the level 2 data, most likely result of \code{simulate_data_l2}.}
}
\description{
Simulate 2-level data with \code{\link[lavaan:simulateData]{lavaan::simulateData}}.
}
