% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{synthethic_data}
\alias{synthethic_data}
\title{Generates Sythethic Data}
\usage{
synthethic_data(within, between, within_cov, between_cov, id, ...)
}
\arguments{
\item{within}{within data that should be retrofit. Most likely output from \code{\link{fs_within}}.}

\item{between}{between data that should be retrofit. Most likely output from \code{\link{fs_between}}.}

\item{within_cov}{the expected cov of within}

\item{between_cov}{the expected cov of between}

\item{...}{one or more cluster identifier}
}
\description{
Transforms factor scores to fit specific covariance matrices and combines them.
}
